Integrate libtopotoolbox (#8)
Fixes #4

This PR attempts to bind libtopotoolbox in a way that works for R and
CRAN following the Writing R Extensions guide.

First, libtopotoolbox is added as a submodule of topotoolboxr. It is
located in the src/libs/ directory. To clone topotoolboxr with the
submodules, use

```
> git clone --recurse-submodules https://github.com/TopoToolbox/topotoolboxr
```

This ensures that the source code of libtopotoolbox is available
within the topotoolboxr repository. The alternative is to insist that
users have libtopotoolbox installed in a system-accessible
location. Since libtopotoolbox is not available in any major system
package manager, I have elected to include the source code.

Because R does not natively support CMake as a build system for
libraries, I have created a POSIX-standard Makefile that simply
compiles all of the C source files in libtopotoolbox/src/ into a
static library. This is all R needs across platforms to be able to
build a libtopotoolbox static library and then link it to the shared
object file that it uses to access C functions from R.

As of now, we must use the TopoToolbox/libtopotoolbox#R-posix-makefile
branch, for which the submodule is currently set up. If this is the
route we decide to go, I will merge that branch into libtopotoolbox
main so we can use it from here.

src/Makevars is used by R to control its build process. It sets up
variables that R needs to find and link the libtopotoolbox library and
then recursively calls make within the libtopotoolbox submodule to
build the static library. R compiles everything in the src/ directory
and links it with the libtopotoolbox static library via the PKG_LIBS
variable.$PKG_CFLAGS is used to define the TOPOTOOLBOX_STATICLIB
macro, which is necessary on Windows to get the right import
declaration. PKG_CPPFLAGS points to the libtopotoolbox include
directory so we can use the libtopotoolbox header file.

I have modified has_topotoolbox to use the corresponding
libtopotoolbox function, which is wrapped here by
`wrap_has_topotoolbox` declared in src/topotoolboxr.h and defined in
topotoolboxr.c.

src/registerDynamicSymbol.c is modified to use the new interface.

R/has_topotoolbox.R uses `wrap_has_topotoolbox` and has been modified
so that it does not have any arguments like the libtopotoolbox
version, and so it returns a Boolean.

tests/testthat/test-has_topotoolbox.R is modified to account for the
new return type of has_topotoolbox

Signed-off-by: William Kearney <william.kearney@uni-potsdam.de>