Install GDAL and PROJ explicitly on macos (#57)
We apparently need to install GDAL and PROJ on macos now. I am not entirely sure why this worked in the past but does not work now (updates to terra or the macos-latest images are the main suspects).

Once we have installed PROJ, we need to set PROJ_DATA to make sure terra can find it. This is a function of the Homebrew install prefix, so we compute it in a separate job step.

See more info about PROJ environment variables at

https://proj.org/en/stable/usage/environmentvars.html

This may need to be communicated to users trying to install topotoolboxr on macos.

Installing stuff from Homebrew causes the macos CI workflow to take a long time. We may need to add some dependency caching.